BasicProject/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.2
exceptiongroup==1.2.1
importlib_metadata==8.0.0
iniconfig==2.0.0
Markdown==3.6
packaging==24.1
pip==22.0.4
pluggy==1.5.0
pytest==8.2.2
setuptools==58.1.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
zipp==3.19.2
